[ADD] CalendarSpaceView 로 Component 분리(#85)
