+++ fairer-iOS/fairer-iOS/Global/Supports/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let rootViewController = UINavigationController(rootViewController: SelectHouseWorkViewController())
+        let rootViewController = UINavigationController(rootViewController: SetHouseWorkViewController())
         window?.rootViewController = rootViewController
         window?.makeKeyAndVisible()
     }
